Feat: restructure the actions, reducers, and thunks files
